+++ src/containers/Tenant/Diagnostics/TopicData/TopicData.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
 
-import {NoSearchResults} from '@gravity-ui/illustrations';
 import {Pagination} from '@gravity-ui/uikit';
-import {skipToken} from '@reduxjs/toolkit/query';
 import {isNil} from 'lodash';
 
 import {DrawerWrapper} from '../../../../components/Drawer';
 import {EmptyFilter} from '../../../../components/EmptyFilter/EmptyFilter';
 import {EnableFullscreenButton} from '../../../../components/EnableFullscreenButton/EnableFullscreenButton';
-import {PageError} from '../../../../components/Errors/PageError/PageError';
 import {Fullscreen} from '../../../../components/Fullscreen/Fullscreen';
 import {
     DEFAULT_TABLE_ROW_HEIGHT,
@@ -17,11 +14,6 @@ import {
 } from '../../../../components/PaginatedTable';
 import {PaginatedTableWithLayout} from '../../../../components/PaginatedTable/PaginatedTableWithLayout';
 import {TableColumnSetup} from '../../../../components/TableColumnSetup/TableColumnSetup';
-import {useClusterWithProxy} from '../../../../store/reducers/cluster/cluster';
-import {partitionsApi} from '../../../../store/reducers/partitions/partitions';
-import {topicApi} from '../../../../store/reducers/topic';
-import type {TopicDataRequest} from '../../../../types/api/topic';
-import {useAutoRefreshInterval} from '../../../../utils/hooks';
 import {useSelectedColumns} from '../../../../utils/hooks/useSelectedColumns';
 import {getIllustration} from '../../../../utils/illustrations';
 import {renderPaginatedTableErrorMessage} from '../../../../utils/renderPaginatedTableErrorMessage';
@@ -37,12 +29,15 @@ import {
 import {generateTopicDataGetter} from './getData';
 import i18n from './i18n';
 import {useTopicDataQueryParams} from './useTopicDataQueryParams';
+import {useTopicPagination} from './useTopicPagination';
+import {useTopicPartitions} from './useTopicPartitions';
+import {useTopicProbeQuery} from './useTopicProbeQuery';
+import {useTopicScroll} from './useTopicScroll';
 import {
     TOPIC_DATA_COLUMNS_TITLES,
     TOPIC_DATA_COLUMNS_WIDTH_LS_KEY,
     TOPIC_DATA_DEFAULT_PAGE_SIZE,
     TOPIC_DATA_FETCH_LIMIT,
-    TOPIC_DATA_MIN_TOTAL_FOR_PAGINATION,
     TOPIC_DATA_SELECTED_COLUMNS_LS_KEY,
     b,
 } from './utils/constants';
@@ -61,20 +56,8 @@ const columns = getAllColumns();
 export function TopicData({scrollContainerRef, path, database, databaseFullPath}: TopicDataProps) {
     const NoSearchResultsImage = getIllustration('NoSearchResults');
 
-    const [autoRefreshInterval] = useAutoRefreshInterval();
-    const useMetaProxy = useClusterWithProxy();
-    const [startOffset, setStartOffset] = React.useState<number>();
-    const [endOffset, setEndOffset] = React.useState<number>();
     const [controlsKey, setControlsKey] = React.useState(0);
-    const [emptyData, setEmptyData] = React.useState(false);
-
-    const [baseOffset, setBaseOffset] = React.useState<number>();
-
-    // Pagination state (1-based, matching Gravity UI Pagination)
-    const [currentPage, setCurrentPage] = React.useState(1);
-
-    // Ref to store pending scroll target after page change
-    const pendingScrollOffset = React.useRef<number | undefined>();
+    const [startOffset, setStartOffset] = React.useState<number>();
 
     const {
         selectedPartition,
@@ -88,21 +71,7 @@ export function TopicData({scrollContainerRef, path, database, databaseFullPath}
         handleActiveOffsetChange,
     } = useTopicDataQueryParams();
 
-    // Scroll to top synchronously before paint when page or partition changes
-    const prevPage = React.useRef(currentPage);
-    const prevPartition = React.useRef(selectedPartition);
-    React.useLayoutEffect(() => {
-        const pageChanged = prevPage.current !== currentPage;
-        const partitionChanged = prevPartition.current !== selectedPartition;
-        if (pageChanged || partitionChanged) {
-            prevPage.current = currentPage;
-            prevPartition.current = selectedPartition;
-            if (scrollContainerRef.current) {
-                scrollContainerRef.current.scrollTop = 0;
-            }
-        }
-    }, [currentPage, selectedPartition, scrollContainerRef]);
-
+    // Clean up topic-specific URL params on unmount
     React.useEffect(() => {
         return () => {
             handleSelectedPartitionChange(undefined);
@@ -111,95 +80,62 @@ export function TopicData({scrollContainerRef, path, database, databaseFullPath}
         };
     }, [handleSelectedPartitionChange, handleSelectedOffsetChange, handleStartTimestampChange]);
 
-    // Total number of offsets in the partition
-    const totalOffsets = React.useMemo(() => {
-        if (isNil(baseOffset) || isNil(endOffset)) {
-            return 0;
-        }
-        return Math.max(endOffset - baseOffset, 0);
-    }, [baseOffset, endOffset]);
-
-    const usePagination = totalOffsets > TOPIC_DATA_MIN_TOTAL_FOR_PAGINATION;
-
-    // Compute pageStartOffset for the current page
-    // When pagination is active: baseOffset + (currentPage - 1) * TOPIC_DATA_DEFAULT_PAGE_SIZE
-    // When pagination is not active: baseOffset (show all offsets)
-    const pageStartOffset = React.useMemo(() => {
-        if (isNil(baseOffset)) {
-            return undefined;
-        }
-        if (!usePagination) {
-            return baseOffset;
-        }
-        return baseOffset + (currentPage - 1) * TOPIC_DATA_DEFAULT_PAGE_SIZE;
-    }, [baseOffset, usePagination, currentPage]);
-
-    const queryParams = React.useMemo(() => {
-        if (isNil(selectedPartition)) {
-            return skipToken;
-        }
-        const params: TopicDataRequest = {database, path, partition: selectedPartition, limit: 1};
-        if (startTimestamp) {
-            params.read_timestamp = startTimestamp;
-        } else if (isNil(selectedOffset)) {
-            return skipToken;
-        } else {
-            params.offset = Number(selectedOffset);
-        }
-        return params;
-    }, [selectedPartition, selectedOffset, startTimestamp, database, path]);
-
-    const {currentData, error, isFetching} = topicApi.useGetTopicDataQuery(queryParams);
+    const {partitions, partitionsLoading, partitionsError, baseOffset, endOffset, setEndOffset} =
+        useTopicPartitions({
+            path,
+            database,
+            databaseFullPath,
+            selectedPartition,
+            onPartitionAutoSelect: handleSelectedPartitionChange,
+        });
 
     const {
-        data: partitions,
-        isLoading: partitionsLoading,
-        error: partitionsError,
-    } = partitionsApi.useGetPartitionsQuery(
-        {path, database, databaseFullPath, useMetaProxy},
-        {pollingInterval: autoRefreshInterval},
+        currentPage,
+        setCurrentPage,
+        totalOffsets,
+        usePagination,
+        pageStartOffset,
+        pageEntitiesCount,
+    } = useTopicPagination({baseOffset, endOffset});
+
+    const setBoundOffsets = React.useCallback(
+        ({
+            startOffset: newStartOffset,
+            endOffset: newEndOffset,
+        }: {
+            startOffset: number;
+            endOffset: number;
+        }) => {
+            setStartOffset(newStartOffset);
+            setEndOffset(newEndOffset);
+        },
+        [setEndOffset],
     );
 
-    const prevSelectedPartition = React.useRef(selectedPartition);
+    const {error, isFetching, emptyData, currentData} = useTopicProbeQuery({
+        path,
+        database,
+        selectedPartition,
+        selectedOffset,
+        startTimestamp,
+        setBoundOffsets,
+    });
 
-    React.useEffect(() => {
-        const selectedPartitionChanged = selectedPartition !== prevSelectedPartition.current;
-        const selectedPartitionData = partitions?.find(
-            ({partitionId}) => partitionId === selectedPartition,
-        );
-        if (selectedPartitionData) {
-            if (isNil(endOffset) || selectedPartitionChanged) {
-                setEndOffset(safeParseNumber(selectedPartitionData.endOffset));
-            }
-            if (isNil(baseOffset) || selectedPartitionChanged) {
-                setBaseOffset(safeParseNumber(selectedPartitionData.startOffset));
-            }
-        }
-        if (selectedPartitionChanged) {
-            prevSelectedPartition.current = selectedPartition;
-        }
-    }, [selectedPartition, partitions, endOffset, baseOffset]);
-
-    React.useEffect(() => {
-        if (partitions && partitions.length && isNil(selectedPartition)) {
-            const firstPartitionId = partitions[0].partitionId;
-            handleSelectedPartitionChange(
-                isNil(firstPartitionId) ? undefined : String(firstPartitionId),
-            );
-        }
-    }, [partitions, selectedPartition, handleSelectedPartitionChange]);
-
-    // Number of entities on the current page
-    const pageEntitiesCount = React.useMemo(() => {
-        if (isNil(pageStartOffset) || isNil(endOffset)) {
-            return 0;
-        }
-        const remaining = endOffset - pageStartOffset;
-        const result = usePagination
-            ? Math.max(Math.min(TOPIC_DATA_DEFAULT_PAGE_SIZE, remaining), 0)
-            : Math.max(remaining, 0);
-        return result;
-    }, [pageStartOffset, endOffset, usePagination]);
+    const {scrollToOffset} = useTopicScroll({
+        scrollContainerRef,
+        baseOffset,
+        endOffset,
+        pageStartOffset,
+        usePagination,
+        currentPage,
+        setCurrentPage,
+        selectedPartition,
+        selectedOffset,
+        startTimestamp,
+        activeOffset,
+        currentData,
+        isFetching,
+    });
 
     const {columnsToShow, columnsToSelect, setColumns} = useSelectedColumns(
         columns,
@@ -207,40 +143,6 @@ export function TopicData({scrollContainerRef, path, database, databaseFullPath}
         TOPIC_DATA_COLUMNS_TITLES,
         DEFAULT_TOPIC_DATA_COLUMNS,
         REQUIRED_TOPIC_DATA_COLUMNS,
-    );
-
-    const setBoundOffsets = React.useCallback(
-        ({startOffset, endOffset}: {startOffset: number; endOffset: number}) => {
-            setStartOffset(startOffset);
-            setEndOffset(endOffset);
-        },
-        [],
-    );
-
-    React.useEffect(() => {
-        //values should be recalculated only when data is fetched
-        if (isFetching || (!currentData && !error)) {
-            return;
-        }
-        const hasMessages = Boolean(currentData?.Messages?.length);
-        setEmptyData(!hasMessages || Boolean(error));
-        if (currentData) {
-            setBoundOffsets({
-                startOffset: safeParseNumber(currentData.StartOffset),
-                endOffset: safeParseNumber(currentData.EndOffset),
-            });
-        }
-    }, [isFetching, currentData, error, setBoundOffsets]);
-
-    const tableFilters = React.useMemo(
-        () => ({
-            path,
-            database,
-            partition: selectedPartition ?? '',
-            isEmpty: emptyData,
-            currentPage,
-        }),
-        [path, database, selectedPartition, emptyData, currentPage],
     );
 
     const resetFilters = React.useCallback(() => {
@@ -257,96 +159,33 @@ export function TopicData({scrollContainerRef, path, database, databaseFullPath}
             resetFilters();
             setCurrentPage(1);
         },
-        [handleSelectedPartitionChange, resetFilters],
+        [handleSelectedPartitionChange, resetFilters, setCurrentPage],
     );
 
-    const scrollToOffset = React.useCallback(
-        (newOffset: number) => {
-            if (isNil(baseOffset) || isNil(endOffset) || isNil(pageStartOffset)) {
-                return;
-            }
-
-            // Guard: offset is out of partition range — nothing to navigate to
-            if (newOffset < baseOffset || newOffset >= endOffset) {
-                return;
-            }
-
-            if (usePagination) {
-                // Calculate which page this offset belongs to (1-based)
-                const absoluteRow = newOffset - baseOffset;
-                const targetPage = Math.floor(absoluteRow / TOPIC_DATA_DEFAULT_PAGE_SIZE) + 1;
-
-                if (targetPage !== currentPage) {
-                    // Need to switch page first, then scroll after render
-                    pendingScrollOffset.current = newOffset;
-                    setCurrentPage(targetPage);
-                    return;
-                }
-            }
-
-            const scrollTop = (newOffset - pageStartOffset) * DEFAULT_TABLE_ROW_HEIGHT;
-            const normalizedScrollTop = Math.max(0, scrollTop);
-            scrollContainerRef.current?.scrollTo({
-                top: normalizedScrollTop,
-                behavior: 'instant',
-            });
-        },
-        [pageStartOffset, baseOffset, endOffset, usePagination, currentPage, scrollContainerRef],
+    const tableFilters = React.useMemo(
+        () => ({
+            path,
+            database,
+            partition: selectedPartition ?? '',
+            isEmpty: emptyData,
+            currentPage,
+        }),
+        [path, database, selectedPartition, emptyData, currentPage],
     );
 
-    // Keep a ref to the latest scrollToOffset to avoid stale closure in useEffect
-    const scrollToOffsetRef = React.useRef(scrollToOffset);
-    React.useLayoutEffect(() => {
-        scrollToOffsetRef.current = scrollToOffset;
-    }, [scrollToOffset]);
+    const getTopicData = React.useMemo(
+        () =>
+            generateTopicDataGetter({
+                setBoundOffsets,
+                baseOffset: pageStartOffset,
+                maxEntities: usePagination ? TOPIC_DATA_DEFAULT_PAGE_SIZE : undefined,
+            }),
+        [pageStartOffset, setBoundOffsets, usePagination],
+    );
 
-    // Handle pending scroll after page change.
-    // Uses scrollToOffsetRef to always call the latest version without adding scrollToOffset to deps.
-    React.useEffect(() => {
-        const pending = pendingScrollOffset.current;
-        if (!isNil(pending)) {
-            pendingScrollOffset.current = undefined;
-            scrollToOffsetRef.current(pending);
-        }
-    }, [currentPage]);
-
-    // On first open: scroll to the offset from URL (selectedOffset or activeOffset).
-    // Consumed once and cleared so subsequent data loads don't re-trigger the scroll.
-    const initialScrollToOffset = React.useRef(selectedOffset ?? activeOffset);
-
-    // When selectedOffset changes, scroll to the first message returned by the API.
-    const shouldScrollToFirstMessage = React.useRef(false);
-    React.useEffect(() => {
-        shouldScrollToFirstMessage.current = true;
-    }, [selectedOffset]);
-
-    React.useEffect(() => {
-        if (isFetching || isNil(baseOffset) || isNil(endOffset)) {
-            return;
-        }
-
-        // Case 1: first open — scroll to the initial offset from URL
-        if (!isNil(initialScrollToOffset.current)) {
-            const targetOffset = Number(initialScrollToOffset.current);
-            initialScrollToOffset.current = undefined;
-            shouldScrollToFirstMessage.current = false;
-            scrollToOffsetRef.current(targetOffset);
-            return;
-        }
-
-        // Case 2: selectedOffset changed — scroll to the first message from API response
-        if (shouldScrollToFirstMessage.current) {
-            shouldScrollToFirstMessage.current = false;
-            const firstMessage = currentData?.Messages?.[0];
-            if (!isNil(firstMessage)) {
-                scrollToOffsetRef.current(safeParseNumber(firstMessage.Offset));
-            }
-        }
-    }, [currentData, isFetching, baseOffset, endOffset]);
-
-    const handlePaginationUpdate = React.useCallback((page: number) => {
-        setCurrentPage(page);
-    }, []);
+    const closeDrawer = React.useCallback(() => {
+        handleActiveOffsetChange(undefined);
+    }, [handleActiveOffsetChange]);
 
     const renderControls = React.useCallback(() => {
         return (
@@ -398,20 +237,6 @@ export function TopicData({scrollContainerRef, path, database, databaseFullPath}
         );
     };
 
-    const getTopicData = React.useMemo(
-        () =>
-            generateTopicDataGetter({
-                setBoundOffsets,
-                baseOffset: pageStartOffset,
-                maxEntities: usePagination ? TOPIC_DATA_DEFAULT_PAGE_SIZE : undefined,
-            }),
-        [pageStartOffset, setBoundOffsets, usePagination],
-    );
-
-    const closeDrawer = React.useCallback(() => {
-        handleActiveOffsetChange(undefined);
-    }, [handleActiveOffsetChange]);
-
     const renderDrawerContent = React.useCallback(() => {
         return (
             <Fullscreen>
@@ -424,9 +249,12 @@ export function TopicData({scrollContainerRef, path, database, databaseFullPath}
         );
     }, [database, path, scrollContainerRef]);
 
-    if (error) {
-        return <PageError error={error} position="left" />;
-    }
+    const handlePaginationUpdate = React.useCallback(
+        (page: number) => {
+            setCurrentPage(page);
+        },
+        [setCurrentPage],
+    );
 
     return (
         !isNil(baseOffset) &&
